Added Data Gathering Script with News API
Added a script to call the News API everything endpoint. It iterates through the articles and writes the content from the API call to a xlsv format. This format can be used by the OpenAI CLI Data Preparation Tool to prepare the data set for fine-tuning a model.

At this time, the script only fills out the article contents as prompts, completions are missing.