Updated gather_news.py and Fixed config_db.py
Updated gather_news.py to run as individual script since wrapping the class was having ImportError issues in config_db.py.

Added a collection release in the config_db.py. If you loaded a collection and wanted to load a new collection, we would need to release the previous collection first. This was an error brought up with Milvus.