Date cleaning & duplicate articles
Added cleaning date function, to convert dates to format such as March 17, 2023. In theory, this would allow the embeddings to pick up on questions that are asking for a specific date more naturally from chat. i.e. What happened in March of 2023, or April 20, 2023 more naturally since people are likely to specify a date by the month rather than the number.

Added a set for keeping track of seen URLs in the NewsAPI call.